moves makeStyles into separate files
