+++ frontend/src/components/PackingList/Item/Item.Styles.js
@@ -1,0 +1,15 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles(() => ({
+    labelUnchecked: {
+        color: "#24305e",
+    },
+    labelChecked: {
+        fontStyle: "italic",
+        textDecoration: "line-through",
+        color: "grey",
+        opacity: "75%",
+    },
+}));
+
+export { useStyles };
