+++ frontend/src/components/PackingList/Item/Item.jsx
@@ -1,21 +1,9 @@
 import React from "react";
+import styles from "./Item.module.css";
+import { useStyles } from "./Item.Styles";
+import { withStyles } from "@material-ui/core";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
 import Checkbox from "@material-ui/core/Checkbox";
-import { makeStyles, withStyles } from "@material-ui/core";
-
-import styles from "./Item.module.css";
-
-const useStyles = makeStyles(() => ({
-    labelUnchecked: {
-        color: "#24305e",
-    },
-    labelChecked: {
-        fontStyle: "italic",
-        textDecoration: "line-through",
-        color: "grey",
-        opacity: "75%",
-    },
-}));
 
 const CustomCheckbox = withStyles({
     root: {
